Move index to Details
