+++ components/details/historyPicker/index.js
@@ -1,0 +1,1 @@
+export * from "./historyPicker";
