+++ components/details/index.js
@@ -1,0 +1,6 @@
+export * from "./documentsTable";
+export * from "./historyPicker";
+export * from "./pdf";
+export * from "./sections";
+export * from "./sectionSeparator";
+export * from "./transactionsTable";
