+++ components/details/pdf/certificatePDF.js
