+++ components/details/pdf/index.js
