+++ components/details/pdf/template.js
