+++ components/details/sectionSeparator/sectionSeparator.js
