+++ components/details/sections/Content.js
@@ -14,12 +14,12 @@ import {
   HistoryTree,
 } from "@licenserocks/kit";
 import styled from "styled-components";
+import Link from "next/link";
 
-import { HistoryPicker } from "components/HistoryPicker";
+import { HistoryPicker } from "components/details/historyPicker";
 import date from "utils/date";
 import { getTransaction } from "utils/ethereum";
 import { withTranslation } from "i18n";
-import Link from "next/link";
 import iconMapper from "./iconMapper";
 
 const HeaderContainer = styled.div`
