+++ components/details/sections/MiningInProgress.js
