+++ components/details/transactionsTable/index.js
@@ -1,0 +1,1 @@
+export * from "./transactionsTable";
