+++ components/index.js
@@ -1,6 +1,1 @@
-export * from "./IndexPage";
-export * from "./PDF";
-export * from "./TransactionsTable";
-export * from "./DocumentsTable";
-export * from "./SectionSeparator";
-export * from "./HistoryPicker";
+export * from "./details";
